탈퇴 약간 수정

+++ SY_project02/frontend/src/hooks/use-authService.js
@@ -125,17 +125,22 @@ export const useDelete = () => {
             data: post,
           })
           .then((res) => {
-            const yes = prompt(`정말 탈퇴를 원하시면 "예" 를 입력 해주세요!`);
-            if (yes === "예") {
-              dispatch(setULogout());
-              dispatch(clearCart());
-              sessionStorage.clear();
-              navigate("/", { replace: true });
-            } else {
-              alert("탈퇴가 취소되었습니다");
+            console.log(res.status, "-");
+
+            if (res.status === 200) {
+              const yes = prompt(`정말 탈퇴를 원하시면 "예" 를 입력 해주세요!`);
+              if (yes === "예") {
+                dispatch(setULogout());
+                dispatch(clearCart());
+                sessionStorage.clear();
+                navigate("/", { replace: true });
+              } else {
+                alert("탈퇴가 취소되었습니다.");
+              }
             }
           })
           .catch((error) => {
+            console.log(error);
             alert("올바른 비밀번호를 입력 해주세요");
             console.error(error);
           });
